[claude] clang-tidy P1: safety-critical fixes — HW verified (0 errors/117K reads)
Fix implicit widening multiplication overflow (uint32 × 2 wraps after ~35 min),
misplaced widening cast, missing default clause, and 27 narrowing conversions
across icm20948, ws2812_status, calibration_manager, and main. Add NOLINTNEXTLINE
for intentional IVP-29 recursion test (RC-1 deviation). Add LL Entry 25: debug
probe vs picotool for iterative flashing — picotool --force corrupts I2C bus.

60s soak: 117,099 IMU reads, 0 errors. Binary: 198,144 bytes (unchanged).

+++ src/calibration/calibration_manager.cpp
@@ -421,7 +421,7 @@ static bool mat_inverse_9x9(const float src[81], float dst[81]) {
         for (uint8_t c = 0; c < 18; c++) {
             aug[col][c] /= pivot;
         }
-        for (int8_t r = col - 1; r >= 0; r--) {
+        for (int8_t r = static_cast<int8_t>(col - 1); r >= 0; r--) {
             float factor = aug[r][col];
             for (uint8_t c = 0; c < 18; c++) {
                 aug[r][c] -= factor * aug[col][c];
